+++ include/ignition/common/detail/PluginMacros.hh
@@ -23,6 +23,7 @@
 #include <typeinfo>
 #include <type_traits>
 #include <unordered_set>
+#include <utility>
 #include "ignition/common/PluginInfo.hh"
 #include "ignition/common/SuppressWarning.hh"
 
@@ -97,7 +98,7 @@ IGN_COMMON_WARN_IGNORE__DELETE_NON_VIRTUAL_DESTRUCTOR \
     } \
     std::unordered_set<std::string> visitedPlugins; \
     ignition::common::PluginInfo * * const ptrToPlugin = \
-        reinterpret_cast<ignition::common::PluginInfo * * const>(_outputInfo); \
+        reinterpret_cast<ignition::common::PluginInfo * *>(_outputInfo); \
     if ( !(*ptrToPlugin) ) \
     { \
        *ptrToPlugin = new ignition::common::PluginInfo; \
@@ -133,9 +134,9 @@ IGN_COMMON_WARN_IGNORE__DELETE_NON_VIRTUAL_DESTRUCTOR \
   /* with each other. */\
   { \
     const bool insertion = visitedPlugins.insert(#pluginName).second; \
-    if (insertion) \
+    if (insertion) /* NOLINT(*) */ \
     { \
-      if (_pluginId == visitedPlugins.size() - 1) \
+      if (_pluginId == visitedPlugins.size() - 1) /* NOLINT(*) */ \
       { \
         /* If the visitedPlugins has reached the requested _pluginId, fill */ \
         /* in the PluginInfo output parameter. */ \
@@ -149,7 +150,7 @@ IGN_COMMON_WARN_IGNORE__DELETE_NON_VIRTUAL_DESTRUCTOR \
       } \
     } \
   \
-    if ( #pluginName == plugin->name ) \
+    if ( #pluginName == plugin->name ) /* NOLINT(*) */ \
     { \
       /* If the name of the desired plugin matches this call to the macro, */ \
       /* add a map entry for the interface specified by this macro. */ \
@@ -163,7 +164,7 @@ IGN_COMMON_WARN_IGNORE__DELETE_NON_VIRTUAL_DESTRUCTOR \
 
 
 #define DETAIL_IGN_COMMON_FINISH_ADDING_PLUGINS \
-    if (_pluginId >= visitedPlugins.size()) \
+    if (_pluginId >= visitedPlugins.size()) /* NOLINT(*) */ \
     { \
       if (plugin) \
         delete plugin; \