+++ src/SystemPaths.cc
@@ -45,8 +45,14 @@ class ignition::common::SystemPathsPrivate
   /// \brief Name of the environment variable to check for plugin paths
   public: std::string pluginPathEnv = "IGN_PLUGIN_PATH";
 
+  /// \brief Name of the environment variable to check for file paths
+  public: std::string filePathEnv = "IGN_FILE_PATH";
+
   /// \brief Paths to plugins
   public: std::list<std::string> pluginPaths;
+
+  /// \brief Paths to files
+  public: std::list<std::string> filePaths;
 
   /// \brief Suffix paths
   public: std::list<std::string> suffixPaths;
@@ -107,6 +113,7 @@ SystemPaths::SystemPaths()
 #ifdef _WIN32
     mkdir(fullPath.c_str());
 #else
+    // cppcheck-suppress ConfigurationNotChecked
     mkdir(fullPath.c_str(), S_IRWXU | S_IRGRP | S_IROTH);
 #endif
   }
@@ -114,6 +121,9 @@ SystemPaths::SystemPaths()
     closedir(dir);
 
   this->dataPtr->logPath = fullPath;
+  // Populate this->dataPtr->filePaths with values from the default
+  // environment variable.
+  this->SetFilePathEnv(this->dataPtr->filePathEnv);
 }
 
 /////////////////////////////////////////////////
@@ -157,7 +167,7 @@ std::string SystemPaths::FindSharedLibrary(const std::string &_libName)
   std::vector<std::string> searchNames =
     this->dataPtr->GenerateLibraryPaths(_libName);
 
-  // TODO return list of paths that match if more than one matches?
+  // TODO(anyone) return list of paths that match if more than one matches?
   for (auto const &possibleName : searchNames)
   {
     if (exists(possibleName))
@@ -169,6 +179,46 @@ std::string SystemPaths::FindSharedLibrary(const std::string &_libName)
   return pathToLibrary;
 }
 
+/////////////////////////////////////////////////
+void SystemPaths::SetFilePathEnv(const std::string &_env)
+{
+  this->dataPtr->filePathEnv = _env;
+  if (!this->dataPtr->filePathEnv.empty())
+  {
+    this->ClearFilePaths();
+    std::string result;
+    if (env(this->dataPtr->filePathEnv, result))
+    {
+      this->AddFilePaths(result);
+    }
+  }
+}
+
+/////////////////////////////////////////////////
+const std::list<std::string> &SystemPaths::FilePaths()
+{
+  return this->dataPtr->filePaths;
+}
+
+/////////////////////////////////////////////////
+void SystemPaths::ClearFilePaths()
+{
+  this->dataPtr->filePaths.clear();
+}
+
+/////////////////////////////////////////////////
+void SystemPaths::AddFilePaths(const std::string &_path)
+{
+  if (_path.size())
+  {
+    std::vector<std::string> paths = Split(_path, Delimiter());
+    for (auto const &path : paths)
+    {
+      std::string normalPath = NormalizeDirectoryPath(path);
+      insertUnique(normalPath, this->dataPtr->filePaths);
+    }
+  }
+}
 /////////////////////////////////////////////////
 std::string SystemPaths::NormalizeDirectoryPath(const std::string &_path)
 {
@@ -245,8 +295,9 @@ std::string SystemPaths::FindFileURI(const std::string &_uri) const
     return this->FindFile(_uri);
   }
 
-  // TODO: Special handling of absolute file:// URIs is needed until the URI
-  //       class is fixed to support absolute URIs
+  // TODO(anyone): Special handling of absolute file:
+  // URIs is needed until the URI
+  // class is fixed to support absolute URIs
   if (common::StartsWith(_uri, "file:///"))
   {
     const auto filename = _uri.substr(std::strlen("file://"));
@@ -312,8 +363,9 @@ std::string SystemPaths::FindFile(const std::string &_filename,
   if (filename.empty())
     return path;
 
-  // TODO: Special handling of absolute file:// URIs is needed until the URI
-  //       class is fixed to support absolute URIs
+  // TODO(anyone): Special handling of absolute file:
+  // URIs is needed until the URI
+  // class is fixed to support absolute URIs
   if (common::StartsWith(filename, "file:///"))
   {
     filename = filename.substr(std::strlen("file://"));
@@ -345,6 +397,20 @@ std::string SystemPaths::FindFile(const std::string &_filename,
     else if (this->dataPtr->findFileCB)
     {
       path = this->dataPtr->findFileCB(filename);
+    }
+  }
+
+  // Look in custom paths.
+  if (path.empty())
+  {
+    for (const std::string &filePath : this->dataPtr->filePaths)
+    {
+      if (exists(joinPaths(filePath, filename)))
+      {
+        path = joinPaths(filePath, filename);
+        ignition::common::replaceAll(path, path, "//", "/");
+        break;
+      }
     }
   }