+++ src/URI.cc
@@ -16,6 +16,7 @@
 */
 
 #include <algorithm>
+#include <cstring>
 #include <list>
 #include <map>
 #include <string>
@@ -26,7 +27,7 @@
 using namespace ignition;
 using namespace common;
 
-static const std::string kSchemeDelim = "://";
+static const char kSchemeDelim[] = "://";
 
 /// \brief URIPath private data.
 class ignition::common::URIPathPrivate
@@ -143,7 +144,8 @@ void URIPath::PushFront(const std::string &_part)
 
   if (part.find('/') != std::string::npos)
   {
-    // TODO: Once URI encoding is implemented, all invalid characters should be
+    // TODO(anyone): Once URI encoding is implemented,
+    // all invalid characters should be
     // encoded, not just slashes.
     ignwarn << "Unencoded slashes in URI part, encoding them." << std::endl;
     part = common::replaceAll(part, "/", "%2F");
@@ -174,7 +176,8 @@ void URIPath::PushBack(const std::string &_part)
 
   if (part.find('/') != std::string::npos)
   {
-    // TODO: Once URI encoding is implemented, all invalid characters should be
+    // TODO(anyone): Once URI encoding is implemented,
+    // all invalid characters should be
     // encoded, not just slashes.
     ignwarn << "Unencoded slashes in URI part, encoding them." << std::endl;
     part = common::replaceAll(part, "/", "%2F");
@@ -258,7 +261,7 @@ bool URIPath::Valid(const std::string &_str)
     return false;
   }
 
-  // TODO: the space should not be
+  // TODO(anyone): the space should not be
   // there, but leaving it out breaks
   // other stuff, e.g. ign-fuel-tools
   // now have URIs with unencoded spaces
@@ -694,7 +697,7 @@ bool URI::Valid(const std::string &_str)
     return false;
   }
 
-  auto from = schemeDelimPos + kSchemeDelim.size();
+  auto from = schemeDelimPos + std::strlen(kSchemeDelim);
   std::string localPath = str.substr(from);
   std::string localQuery;
   std::string localFragment;
@@ -731,7 +734,7 @@ bool URI::Parse(const std::string &_str)
     return false;
 
   auto schemeDelimPos = _str.find(kSchemeDelim);
-  auto from = schemeDelimPos + kSchemeDelim.size();
+  auto from = schemeDelimPos + std::strlen(kSchemeDelim);
   std::string localScheme = _str.substr(0, schemeDelimPos);
   std::string localPath = _str.substr(from);
   std::string localQuery;