+++ src/URI_TEST.cc
@@ -36,20 +36,20 @@ TEST(URITEST, URIPath)
 
   path1.PushFront("part 0");
   EXPECT_EQ(path1.Str(), "part 0/part1/part2");
-  // TODO: switch to the following once URI class is upgraded
+  // TODO(anyone): switch to the following once URI class is upgraded
   // path1.PushFront("part%200");
   // EXPECT_EQ(path1.Str(), "part%200/part1/part2");
   EXPECT_FALSE(path1.IsAbsolute());
 
   path2 = path1 / "part3";
   EXPECT_EQ(path2.Str(), "part 0/part1/part2/part3");
-  // TODO: switch to the following once URI class is upgraded
+  // TODO(anyone): switch to the following once URI class is upgraded
   // EXPECT_EQ(path2.Str(), "part%200/part1/part2/part3");
   EXPECT_FALSE(path2.IsAbsolute());
 
   path1 /= "part3";
   EXPECT_EQ(path1.Str(), "part 0/part1/part2/part3");
-  // TODO: switch to the following once URI class is upgraded
+  // TODO(anyone): switch to the following once URI class is upgraded
   // EXPECT_EQ(path1.Str(), "part%200/part1/part2/part3");
   EXPECT_FALSE(path1.IsAbsolute());
 
@@ -190,7 +190,7 @@ TEST(URITEST, URIPathString)
   EXPECT_TRUE(URIPath::Valid("/part1/part2/"));
 
   EXPECT_TRUE(URIPath::Valid("/part 1/part 2/"));
-  // TODO: switch to the following once URI class is upgraded
+  // TODO(anyone): switch to the following once URI class is upgraded
   // EXPECT_FALSE(URIPath::Valid("/part 1/part 2/"));
   // EXPECT_TRUE(URIPath::Valid("/part+1/part+2"));
   // EXPECT_TRUE(URIPath::Valid("/part%201/part%202"));
@@ -213,7 +213,7 @@ TEST(URITEST, URIPathString)
   EXPECT_TRUE(path.Parse("/part1/part2/"));
 
   EXPECT_TRUE(path.Parse("/part 1/part 2/"));
-  // TODO: switch to the following once URI class is upgraded
+  // TODO(anyone): switch to the following once URI class is upgraded
   // EXPECT_FALSE(path.Parse("/part 1/part 2/"));
   // EXPECT_TRUE(path.Parse("/part+1/part+2"));
   // EXPECT_TRUE(path.Parse("/part%201/part%202"));
@@ -236,7 +236,7 @@ TEST(URITEST, URIPathString)
   EXPECT_NO_THROW(URIPath("/part1/part2"));
   EXPECT_NO_THROW(URIPath("/part 1/part2/"));
 
-  // TODO: uncomment following once URI class is upgraded
+  // TODO(anyone): uncomment following once URI class is upgraded
   // EXPECT_NO_THROW(URIPath("/part+1/part+2"));
   // EXPECT_NO_THROW(URIPath("/part%201/part%202"));
 }
@@ -382,7 +382,7 @@ TEST(URITEST, URIFragmentString)
   EXPECT_TRUE(URIFragment::Valid(""));
   EXPECT_TRUE(URIFragment::Valid("#fragment"));
 
-  // TODO: uncomment following once URI class is upgraded
+  // TODO(anyone): uncomment following once URI class is upgraded
   // EXPECT_TRUE(URIFragment::Valid("#fragment/?!$&'()*+,;=:@%20fragment"));
 
   EXPECT_FALSE(fragment.Parse("##"));
@@ -398,7 +398,7 @@ TEST(URITEST, URIFragmentString)
   EXPECT_TRUE(fragment.Parse(""));
   EXPECT_TRUE(fragment.Parse("#fragment"));
 
-  // TODO: uncomment following once URI class is upgraded
+  // TODO(anyone): uncomment following once URI class is upgraded
   // EXPECT_TRUE(fragment.Parse("#fragment/?!$&'()*+,;=:@%20FRAGMENT"));
 
   // it should still be valid
@@ -583,7 +583,7 @@ TEST(URITEST, URIString)
   EXPECT_TRUE(URI::Valid("scheme://part1 /part2 ?k1=v1&k2=v2"));
   EXPECT_TRUE(URI::Valid("scheme://part  1  /part  2  ?k1=v1&k2=v2"));
   EXPECT_TRUE(URI::Valid("scheme with space://part 1/part 2?k1=v1&k2=v2"));
-  // TODO: switch to the following once URI class is upgraded
+  // TODO(anyone): switch to the following once URI class is upgraded
   // EXPECT_FALSE(URI::Valid("scheme://part 1/part 2?k1=v1&k2=v2"));
   // EXPECT_FALSE(URI::Valid("scheme://part1 /part2 ?k1=v1&k2=v2"));
   // EXPECT_FALSE(URI::Valid("scheme://part  1  /part  2  ?k1=v1&k2=v2"));